+++ casper/java/ast.h
@@ -34,13 +34,13 @@ namespace casper
     {
         class FakeJavaParser;
         class FakeJavaExpression;
-        
+
         class Ast
         {
-            
+
             friend class FakeJavaParser;
             friend class FakeJavaExpression;
-            
+
         protected: // Data
 
             std::vector<AstNode*> allocated_nodes_;
@@ -62,7 +62,7 @@ namespace casper
             AstNode* Bool       (bool a_bool);
 
             AstNode* NewAstNode ();
-            AstNode* NewAstNode (const int a_num);
+            AstNode* NewAstNode (const double a_num);
             AstNode* NewAstNode (const std::string& a_text);
             AstNode* NewAstNode (const casper::java::AstNode::Type a_type, const std::string& a_text);
